Add authors
This is the tentative list that Ralf, Stefan, and I agreed on.
We went back to ancient history (`git shortlog -sn v0.9.2..HEAD`), and
tried to weigh merge commits (indicative of review effort) and other
contributions like release management, docs, mentoring, and so on.

It is open to discussion, but we think it is fairly reasonable.